+++ test/client_test.go
@@ -339,7 +339,7 @@ func waitUntilEndpointSynchronized(ctx context.Context, c driver.Client, dbname 
 func TestCreateClientHttpConnection(t *testing.T) {
 	conn, err := http.NewConnection(http.ConnectionConfig{
 		Endpoints: getEndpointsFromEnv(t),
-		TLSConfig: &tls.Config{InsecureSkipVerify: true},
+		Transport: NewConnectionTransport(),
 	})
 	if err != nil {
 		t.Fatalf("Failed to create new http connection: %s", describe(err))
@@ -350,37 +350,6 @@ func TestCreateClientHttpConnection(t *testing.T) {
 	})
 	if err != nil {
 		t.Fatalf("Failed to create new client: %s", describe(err))
-	}
-}
-
-// TestCreateClientHttpConnectionCustomTransport creates an HTTP connection to the environment specified
-// endpoints with a custom HTTP roundtripper and creates a client for that.
-func TestCreateClientHttpConnectionCustomTransport(t *testing.T) {
-	conn, err := http.NewConnection(http.ConnectionConfig{
-		Endpoints: getEndpointsFromEnv(t),
-		Transport: &httplib.Transport{},
-		TLSConfig: &tls.Config{InsecureSkipVerify: true},
-	})
-	if err != nil {
-		t.Fatalf("Failed to create new http connection: %s", describe(err))
-	}
-	c, err := driver.NewClient(driver.ClientConfig{
-		Connection:     conn,
-		Authentication: createAuthenticationFromEnv(t),
-	})
-	if err != nil {
-		t.Fatalf("Failed to create new client: %s", describe(err))
-	}
-	timeout := 3 * time.Minute
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	defer cancel()
-	if up := waitUntilServerAvailable(ctx, c, t); up != nil {
-		t.Fatalf("Connection is not available in %s: %s", timeout, describe(up))
-	}
-	if info, err := c.Version(driver.WithDetails(ctx)); err != nil {
-		t.Errorf("Version failed: %s", describe(err))
-	} else {
-		t.Logf("Got server version %s", info)
 	}
 }